Serialized course data
Add a method to the scraper to print out course data as a serialized array.
Add a text file storing the output of the parse run.
Add a method to load the serialized text file and print out the array contents.
Add a link on the welcome page to call the print method.